fix conflict issues of the consensus state and safe maps
